chore(origin-ui-certificate-data): improve folder structure
