+++ packages/ui/libs/certificate/data/src/cached/cachedFuelTypes.ts
