+++ packages/ui/libs/certificate/data/src/fetching/allDevices.ts
@@ -1,6 +1,6 @@
 import { useDeviceRegistryControllerGetAll } from '@energyweb/origin-device-registry-api-react-query-client';
 import { useDeviceControllerGetAll } from '@energyweb/origin-device-registry-irec-local-api-react-query-client';
-import { composePublicDevices } from './utils';
+import { composePublicDevices } from '../utils';
 
 export const useApiAllDevices = () => {
   const { data: allOriginDevices, isLoading: isOriginDevicesLoading } =
