+++ packages/ui/libs/certificate/data/src/fetching/fuelTypes.ts
