+++ packages/ui/libs/certificate/data/src/fetching/myDevices.ts
@@ -1,7 +1,7 @@
 import { useDeviceRegistryControllerGetMyDevices } from '@energyweb/origin-device-registry-api-react-query-client';
 import { useDeviceControllerGetMyDevices } from '@energyweb/origin-device-registry-irec-local-api-react-query-client';
-import { composeMyDevices } from './utils';
-import { ComposedDevice } from './types';
+import { composeMyDevices } from '../utils';
+import { ComposedDevice } from '../types';
 
 export const useApiMyDevices = () => {
   const { data: allOriginDevices, isLoading: isOriginDevicesLoading } =
