+++ packages/ui/libs/certificate/data/src/handlers/depositCertificate.ts
@@ -1,5 +1,4 @@
 import { useAccountControllerGetAccount } from '@energyweb/exchange-react-query-client';
-import { IClaimData } from '@energyweb/issuer';
 import {
   CertificateDTO,
   getIrecCertificateControllerGetAllQueryKey,
@@ -12,18 +11,21 @@ import { PowerFormatter } from '@energyweb/origin-ui-utils';
 import { BigNumber } from 'ethers';
 import { useTranslation } from 'react-i18next';
 import { useQueryClient } from 'react-query';
-import { useGetBlockchainCertificateHandler } from './getBlockchainCertificate';
+import { useGetBlockchainCertificateHandler } from '../fetching';
 
-export const useRetireCertificateHandler = (resetList: () => void) => {
+export const useDepositCertificateHandler = (resetList: () => void) => {
   const { t } = useTranslation();
   const queryClient = useQueryClient();
   const blockchainCertificatesQueryKey =
     getIrecCertificateControllerGetAllQueryKey();
 
+  const { data: account, isLoading: isAccountLoading } =
+    useAccountControllerGetAccount();
+
   const { getBlockchainCertificate, isLoading: isGetBlockchainLoading } =
     useGetBlockchainCertificateHandler();
 
-  const retireHandler = async <Id>(id: Id, amount: string) => {
+  const depositHandler = async <Id>(id: Id, amount: string) => {
     try {
       const onChainCertificate = await getBlockchainCertificate(
         id as unknown as CertificateDTO['id']
@@ -31,10 +33,8 @@ export const useRetireCertificateHandler = (resetList: () => void) => {
       const formattedAmount = BigNumber.from(
         PowerFormatter.getBaseValueFromValueInDisplayUnit(Number(amount))
       );
-      //mock
-      const claimData: IClaimData = {};
-      const transaction = await onChainCertificate.claim(
-        claimData,
+      const transaction = await onChainCertificate.transfer(
+        account.address,
         formattedAmount
       );
       const receipt = await transaction.wait();
@@ -56,7 +56,7 @@ export const useRetireCertificateHandler = (resetList: () => void) => {
     }
   };
 
-  const isLoading = isGetBlockchainLoading;
+  const isLoading = isAccountLoading || isGetBlockchainLoading;
 
-  return { retireHandler, isLoading };
+  return { depositHandler, isLoading };
 };
