+++ packages/ui/libs/certificate/data/src/handlers/retireCertificate.ts
@@ -1,4 +1,4 @@
-import { useAccountControllerGetAccount } from '@energyweb/exchange-react-query-client';
+import { IClaimData } from '@energyweb/issuer';
 import {
   CertificateDTO,
   getIrecCertificateControllerGetAllQueryKey,
@@ -11,21 +11,18 @@ import { PowerFormatter } from '@energyweb/origin-ui-utils';
 import { BigNumber } from 'ethers';
 import { useTranslation } from 'react-i18next';
 import { useQueryClient } from 'react-query';
-import { useGetBlockchainCertificateHandler } from './getBlockchainCertificate';
+import { useGetBlockchainCertificateHandler } from '../fetching';
 
-export const useDepositCertificateHandler = (resetList: () => void) => {
+export const useRetireCertificateHandler = (resetList: () => void) => {
   const { t } = useTranslation();
   const queryClient = useQueryClient();
   const blockchainCertificatesQueryKey =
     getIrecCertificateControllerGetAllQueryKey();
 
-  const { data: account, isLoading: isAccountLoading } =
-    useAccountControllerGetAccount();
-
   const { getBlockchainCertificate, isLoading: isGetBlockchainLoading } =
     useGetBlockchainCertificateHandler();
 
-  const depositHandler = async <Id>(id: Id, amount: string) => {
+  const retireHandler = async <Id>(id: Id, amount: string) => {
     try {
       const onChainCertificate = await getBlockchainCertificate(
         id as unknown as CertificateDTO['id']
@@ -33,8 +30,10 @@ export const useDepositCertificateHandler = (resetList: () => void) => {
       const formattedAmount = BigNumber.from(
         PowerFormatter.getBaseValueFromValueInDisplayUnit(Number(amount))
       );
-      const transaction = await onChainCertificate.transfer(
-        account.address,
+      //mock
+      const claimData: IClaimData = {};
+      const transaction = await onChainCertificate.claim(
+        claimData,
         formattedAmount
       );
       const receipt = await transaction.wait();
@@ -56,7 +55,7 @@ export const useDepositCertificateHandler = (resetList: () => void) => {
     }
   };
 
-  const isLoading = isAccountLoading || isGetBlockchainLoading;
+  const isLoading = isGetBlockchainLoading;
 
-  return { depositHandler, isLoading };
+  return { retireHandler, isLoading };
 };
