+++ packages/ui/libs/certificate/data/src/index.ts
@@ -1,21 +1,7 @@
-export * from './exchangeCertificates';
-export * from './certificationRequests';
-export * from './pendingCertificationRequests';
-export * from './fileUDownload';
-export * from './fuelTypes';
-export * from './certificates';
-export * from './allDevices';
-export * from './myDevices';
+export * from './cached';
+export * from './fetching';
+export * from './handlers';
+
 export * from './utils';
 export * from './types';
 export * from './mock';
-export * from './cachedExchangeCertificates';
-export * from './cachedMyDevices';
-export * from './cachedFuelTypes';
-export * from './sellCertificate';
-export * from './withdrawCertificate';
-export * from './cachedUser';
-export * from './blockchainCertificates';
-export * from './cachedBlockchainCertificates';
-export * from './depositCertificate';
-export * from './retireCertificate';
